More Phase 3 changes
FEMR49/159: Require a password match when creating a new user
https://teamfemr.atlassian.net/browse/FEMR-159

Required password confirmation for creating a user.
Added logic to UsersController.java for checking if passwords match.
Added required field in create.scala.html for a matching password to be
entered.
Added getters and setters to CreateViewModel.java to be used with logic
in UsersController.java